+++ crates/bee_orm_macro/src/lib.rs
@@ -1,7 +1,7 @@
 // Copyright (c) 2026 erik <erik@erik.xyz> — https://erik.xyz
 use proc_macro::TokenStream;
 use quote::quote;
-use syn::{parse_macro_input, DeriveInput};
+use syn::{DeriveInput, parse_macro_input};
 
 #[proc_macro_derive(Model, attributes(bee))]
 pub fn derive_model(input: TokenStream) -> TokenStream {